samantha_ai_assistant/apps/samantha-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pydub
starlette
# Add other dependencies as needed, e.g., for Whisper, ElevenLabs
# openai-whisper
# elevenlabs